backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
PyJWT
httpx
openai
